Upgrade Gold.ConsoleMenu to target .NET Standard 2.0
This will work with .NET Framework 4.6.1 and above, and with .NET Core.

To use Gold.Gold.ConsoleMenu with earlier versions of the .NET Framework
continue to use the Gold.ConsoleMenu 2.1.0 NuGet package.  That is
identical to this version 3.1.1 but is built against .NET Framework 2.0.

The target framework for the MenuDemo project has also been updated, so
it will work with the new version of Gold.ConsoleMenu.